pyproject.toml
[tool.poetry]
name = "yt-etl-project-api-extract"
version = "0.1.0"
description = "Api Extract"
authors = ["Gustavo Souza"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.1"
requests = "^2.32.3"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
